requirements.txt
django==5.0.4
tensorflow==2.18.0
keras==3.8.0
numpy==1.26.0
pillow==10.3.0
matplotlib==3.8.0
h5py==3.11.0